+++ libs/ngx.command/src/command.directive.ts
@@ -9,8 +9,7 @@ import {
 	ChangeDetectorRef,
 	inject,
 } from "@angular/core";
-import { Subject } from "rxjs";
-import { tap, delay, takeUntil } from "rxjs/operators";
+import { Subject, tap, delay, takeUntil } from "rxjs";
 
 import { type CommandOptions, COMMAND_OPTIONS } from "./command.options";
 import { Command } from "./command";
@@ -100,7 +99,7 @@ export class CommandDirective implements OnInit, OnDestroy {
 	private _destroy$ = new Subject<void>();
 
 	ngOnInit(): void {
-		// console.log("[ssvCommand::init]", this.config);
+		// console.log("[ssvCommand::init]", this.globalOptions);
 		if (!this.commandOrCreator) {
 			throw new Error(`${NAME_CAMEL}: [${NAME_CAMEL}] should be defined!`);
 		} else if (isCommand(this.commandOrCreator)) {