+++ libs/ngx.command/src/command.ts
@@ -1,7 +1,45 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { Observable, combineLatest, Subscription, Subject, BehaviorSubject, of, EMPTY } from "rxjs";
-import { tap, map, filter, switchMap, catchError, finalize, take } from "rxjs/operators";
-import { ICommand } from "./command.model";
+import {
+	Observable, combineLatest, Subscription, Subject, BehaviorSubject, of, EMPTY,
+	tap, map, filter, switchMap, catchError, finalize, take,
+ } from "rxjs";
+import type { ICommand } from "./command.model";
+import { DestroyRef, inject, isSignal, type Signal } from "@angular/core";
+import { toObservable } from "@angular/core/rxjs-interop";
+
+export type ExecuteFn = (...args: any[]) => unknown;
+export type ExecuteAsyncFn = (...args: any[]) => Observable<unknown> | Promise<unknown>;
+export type CanExecute = Signal<boolean> | Observable<boolean>;
+
+/** Creates an async {@link Command}. Must be used within an injection context.
+ * NOTE: this auto injects `DestroyRef` and handles auto destroy. {@link ICommand.autoDestroy} should not be used.
+ */
+export function createCommandAsync(
+	execute: ExecuteAsyncFn,
+	canExecute$?: CanExecute,
+): Command {
+	return createCommand(execute, canExecute$, true);
+}
+
+/** Creates a {@link Command}. Must be used within an injection context.
+ * NOTE: this auto injects `DestroyRef` and handles auto destroy. {@link ICommand.autoDestroy} should not be used.
+ */
+export function createCommand(
+	execute: ExecuteFn,
+	canExecute$?: CanExecute,
+	isAsync?: boolean,
+): Command {
+	const destroyRef = inject(DestroyRef);
+
+	const cmd = new Command(execute, canExecute$, isAsync);
+	cmd.autoDestroy = false;
+
+	destroyRef.onDestroy(() => {
+		// console.warn("[createCommandAsync::destroy]");
+		cmd.destroy();
+	});
+	return cmd;
+}
 
 /**
  * Command object used to encapsulate information which is needed to perform an action.
@@ -46,14 +84,14 @@ export class Command implements ICommand {
 	 * @param isAsync Indicates that the execute function is async e.g. Observable.
 	 */
 	constructor(
-		execute: (...args: unknown[]) => unknown,
-		canExecute$?: Observable<boolean>,
+		execute: ExecuteFn,
+		canExecute$?: CanExecute,
 		isAsync?: boolean,
 	) {
 		if (canExecute$) {
 			this.canExecute$ = combineLatest([
 				this._isExecuting$,
-				canExecute$
+				isSignal(canExecute$) ? toObservable(canExecute$) : canExecute$
 			]).pipe(
 				map(([isExecuting, canExecuteResult]) => {
 					// console.log("[command::combineLatest$] update!", { isExecuting, canExecuteResult });
@@ -161,8 +199,8 @@ export class Command implements ICommand {
 export class CommandAsync extends Command {
 
 	constructor(
-		execute: (...args: any[]) => Observable<unknown> | Promise<unknown>,
-		canExecute$?: Observable<boolean>,
+		execute: ExecuteAsyncFn,
+		canExecute$?: CanExecute,
 	) {
 		super(execute, canExecute$, true);
 	}